If more than 1 exercise in a feature, enable try another and only create 1 exercise in the assignment
